+++ cocos/2d/CCFontFNT.cpp
@@ -149,7 +149,7 @@ FontAtlas * FontFNT::createFontAtlas()
     if (!_configuration->_fontDefDictionary)
         return nullptr;
     
-    int numGlyphs = _configuration->_characterSet->size();
+    long numGlyphs = _configuration->_characterSet->size();
     if (!numGlyphs)
         return nullptr;
