+++ cocos/2d/CCFontFreeType.h
@@ -46,8 +46,8 @@ public:
     virtual Size        * getAdvancesForTextUTF16(unsigned short *text, int &outNumLetters) const override;
     virtual GlyphDef    * getGlyphDefintionsForText(const char *text, int &outNumGlyphs,    bool UTF16text = false) const override;
     unsigned char       * getGlyphBitmap(unsigned short theChar, int &outWidth, int &outHeight) const override;
-    virtual int           getFontMaxHeight() const override;
-    virtual int           getLetterPadding() const override;
+    virtual long           getFontMaxHeight() const override;
+    virtual long           getLetterPadding() const override;
     
     bool getBBOXFotChar(unsigned short theChar, Rect &outRect) const;
 
@@ -64,9 +64,9 @@ private:
     bool initFreeType();
     FT_Library getFTLibrary();
     
-    int  getAdvanceForChar(unsigned short theChar) const;
-    int  getBearingXForChar(unsigned short theChar) const;
-    int  getHorizontalKerningForChars(unsigned short firstChar, unsigned short secondChar) const;
+    long  getAdvanceForChar(unsigned short theChar) const;
+    long  getBearingXForChar(unsigned short theChar) const;
+    long  getHorizontalKerningForChars(unsigned short firstChar, unsigned short secondChar) const;
     
     static FT_Library _FTlibrary;
     static bool       _FTInitialized;
